feat: Add show user's name on Home component
