+++ src/components/Navbar/Navbar.jsx
@@ -11,15 +11,10 @@ const Navbar = () => {
 				<img src="../../src/assets/logo.png" alt="Logo CoTravel" />
 				<h1>CoTravel</h1>
 			</div>
-			{/* <button className={styles.hamburgerBtn} type="button">
-				<div className={styles.hamburgerBtnBox}>
-					MENU
-				</div>
-			</button> */}
+
 			<button className={styles.logoutBtn} onClick={logout} type="button">
 				Wyloguj się
 			</button>
-			{/* <nav className={styles.nav}></nav> */}
 		</div>
 	);
 };
